refactor: move transaction components
