+++ frontend/src/lib/components/transaction/TransactionQRCode.svelte
