+++ frontend/src/lib/modals/neurons/IncreaseNeuronStakeModal.svelte
@@ -3,7 +3,7 @@
   import { topUpNeuron } from "$lib/services/neurons.services";
   import type { NewTransaction } from "$lib/types/transaction";
   import { toastsSuccess } from "$lib/stores/toasts.store";
-  import TransactionModal from "../accounts/NewTransaction/TransactionModal.svelte";
+  import TransactionModal from "$lib/components/transaction/TransactionModal.svelte";
   import { startBusy, stopBusy } from "$lib/stores/busy.store";
   import { isAccountHardwareWallet } from "$lib/utils/accounts.utils";
   import { createEventDispatcher, onMount } from "svelte";
