+++ frontend/src/lib/modals/sns/neurons/SnsTransactionModal.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import TransactionModal from "$lib/modals/accounts/NewTransaction/TransactionModal.svelte";
+  import TransactionModal from "$lib/components/transaction/TransactionModal.svelte";
   import type { Principal } from "@dfinity/principal";
   import type { WizardStep } from "@dfinity/gix-components";
   import type { Token, TokenAmount } from "@dfinity/nns";
